:recycle: issue #20 フォルダ構成を整理し、`RealmDataBaseManager`をインフラ層に定義
